fix: compilation issue and some lint from checkpatch add a .gitignore add new signing section in Readme

+++ hid-ft260.c
@@ -1273,7 +1273,7 @@ static int ft260_gpio_init(struct ft260_device *dev,
 {
 	int ret;
 	int label_sz;
-	char * label;
+	char *label;
 	struct ft260_get_chip_version_report ver;
 	struct hid_device *hdev = dev->hdev;
 	char prefix[] = "ft260_";
@@ -1807,7 +1807,7 @@ static int ft260_uart_receive_chars(struct ft260_device *port, u8 *data, u8 leng
 	return ret;
 }
 
-static int ft260_uart_write(struct tty_struct *tty, const u8 *buf, int cnt)
+static ssize_t ft260_uart_write(struct tty_struct *tty, const u8 *buf, unsigned long cnt)
 {
 	struct ft260_device *port = tty->driver_data;
 	int len, ret, diff;
